requirements-rocm.txt
tensorflow-rocm==2.2.0
numpy==1.20.0
Pillow==9.4.0
scipy==1.4.1
protobuf==3.20.*